requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Web framework
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-cors>=4.0.0

# Data processing and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Natural Language Processing
spacy>=3.6.0
openai>=1.0.0

# API and data fetching
requests>=2.31.0
kaggle>=1.5.0

# Development tools
jupyter>=1.0.0
python-dotenv>=1.0.0
pytest>=7.4.0

# Database migrations
alembic>=1.11.0

# Data visualization (web)
plotly>=5.15.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.65.0
